Separate logical plan verification from binding resolution
Move reusable verification into LogicalPlanVerifier, simplify extension opt-in to a nullable stable identity, and keep ordinary column binding resolution unchanged. Tighten stable fact validation without enumerating every accepted logical type and remove unnecessary enum storage qualifiers.
