+++ src/include/duckdb/execution/column_binding_resolver.hpp
@@ -10,12 +10,9 @@
 
 #include "duckdb/planner/logical_operator_visitor.hpp"
 #include "duckdb/planner/column_binding_map.hpp"
-#include "duckdb/planner/logical_plan_compiler_result.hpp"
 #include "duckdb/common/vector.hpp"
 
 namespace duckdb {
-
-struct ColumnBindingVerificationState;
 
 //! The ColumnBindingResolver resolves ColumnBindings into base tables
 //! (table_index, column_index) into physical indices into the DataChunks that
@@ -25,20 +22,12 @@ public:
 	explicit ColumnBindingResolver(bool verify_only = false);
 
 	void VisitOperator(LogicalOperator &op) override;
-	static void Verify(ClientContext &context, LogicalOperator &op);
-	DUCKDB_API static LogicalPlanCompilerResult<LogicalPlanVerificationSuccess> VerifyAlways(LogicalOperator &op);
 
 protected:
 	vector<ColumnBinding> bindings;
 	vector<LogicalType> types;
 	bool verify_only;
-	optional_ptr<ColumnBindingVerificationState> verification_state;
 
 	unique_ptr<Expression> VisitReplace(BoundColumnRefExpression &expr, unique_ptr<Expression> *expr_ptr) override;
-	explicit ColumnBindingResolver(ColumnBindingVerificationState &verification_state);
-	static bool ResolveOperatorTypes(LogicalOperator &op, ColumnBindingVerificationState &verification_state);
-	static void VerifyColumnBindings(LogicalOperator &op, ColumnBindingVerificationState &verification_state);
-	static LogicalPlanCompilerResult<LogicalPlanVerificationSuccess>
-	VerifyAlwaysInternal(LogicalOperator &op, optional_ptr<string> first_error);
 };
 } // namespace duckdb
