+++ src/include/duckdb/planner/logical_plan_verifier.hpp
@@ -1,0 +1,36 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb/planner/logical_plan_verifier.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include "duckdb/execution/column_binding_resolver.hpp"
+#include "duckdb/planner/logical_plan_compiler_result.hpp"
+
+namespace duckdb {
+
+struct LogicalPlanVerificationState;
+
+class LogicalPlanVerifier : private ColumnBindingResolver {
+public:
+	static void Verify(ClientContext &context, LogicalOperator &op);
+	DUCKDB_API static LogicalPlanCompilerResult<LogicalPlanVerificationSuccess> VerifyAlways(LogicalOperator &op);
+
+private:
+	explicit LogicalPlanVerifier(LogicalPlanVerificationState &verification_state);
+
+	unique_ptr<Expression> VisitReplace(BoundColumnRefExpression &expr, unique_ptr<Expression> *expr_ptr) override;
+	static bool ResolveOperatorTypes(LogicalOperator &op, LogicalPlanVerificationState &verification_state);
+	static void VerifyColumnBindings(LogicalOperator &op, LogicalPlanVerificationState &verification_state);
+	static LogicalPlanCompilerResult<LogicalPlanVerificationSuccess>
+	VerifyAlwaysInternal(LogicalOperator &op, optional_ptr<string> first_error);
+
+private:
+	LogicalPlanVerificationState &verification_state;
+};
+
+} // namespace duckdb
